[expo-router] add context scoped route info
