+++ packages/expo-router/src/hooks/useCurrentRouteInfo.ts
@@ -1,9 +1,8 @@
 'use client';
 
-import { useMemo } from 'react';
+import { use } from 'react';
 
-import { getCachedRouteInfo } from '../global-state/routeInfoCache';
-import { useStateForPath } from '../react-navigation/native';
+import { RouteInfoContext } from '../global-state/RouteInfoContext';
 
 /**
  * Returns route info for a screen it is called from.
@@ -11,7 +10,5 @@ import { useStateForPath } from '../react-navigation/native';
  * @experimental
  */
 export function useCurrentRouteInfo() {
-  const state = useStateForPath();
-  const routeInfo = useMemo(() => (state ? getCachedRouteInfo(state) : undefined), [state]);
-  return routeInfo;
+  return use(RouteInfoContext);
 }
